% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomica.R
\name{genomica}
\alias{genomica}
\title{Genomica - LMM driven (FDR adjusted) KEGG Orthologs enrichment analysis}
\usage{
genomica(
  Data,
  Metadata,
  Predictors,
  P1_Levels = c(0),
  P2_Levels = c(0),
  R_Effects,
  R1_Levels,
  Already_Log10_transformed = c("NO"),
  Folder_Name
)
}
\arguments{
\item{Data}{A data frame with samples as columns and features as rows.}

\item{Metadata}{A data frame with samples as rows.}

\item{Predictors}{example: c('Treatment','Growing_Phase') - This vector will be used to specify the predictors independent variables, according to your study design. Currently, Genomica accepts maximum two predictors, in which case it will automatically calculate the interaction between the two according to the formula predictor 1 * predictor 2 within the linear mixed model.}

\item{R_Effects}{example: c('Block') - This vector defines the random effects according to your study design.}

\item{R1_Levels}{example: c('Block1,'Block2,Block3,Block4,'Block5,Block6,Block7) - This vector will specify the levels for the random effect, according to your study design.}

\item{Already_Log10_transformed}{Default = c('NO'). If c('YES'), the log10 (n+1) transformation of the features in the object Data will not be performed.}

\item{Folder_Name}{Example: c('Test') - This string will be attached to the folder name in which the results will be generated (e.g.,"Genomica_Output_Test/")}

\item{P1_levels}{example: c('Control','Treatment1','Treatment2) - If your main predictor is not a numerical variable, this will set the levels for your main predictor (the model will test the comparisons towards the first term in this vector, such as Treatment1 Vs Control and Treatment2 Vs Control in the example).}

\item{P2_levels}{example: c('Starter','Grower') - If your second predictor is not a numerical variable, this will set the levels for your main predictor (the model will test the comparisons towards the first term in this vector, such as Grower Vs Starter in the example).}
}
\description{
This function will perform a linear mixed model on all the features in your data set, whilst also taking into account of your experimental design (i.e., fixed and random effects). Thus, Genomica will carry out an enrichment analysis of the orthologs based on the given predictor list. Please see the vignette for more information on how to prepare your data frames for the analysis.
}
\examples{
#Load Data data frame (Data_Demo)
Data<-Data_Demo
#In this example Data is a data frame of 500 KEGG orthologs (KOs) in rows, across 35 samples (columns)
print(Data[,1:5])
Metadata<-Metadata_Demo
#in this example Metadata is a data frame with 35 rows (sample IDs) and 5 columns (SampleID, Treatment and Block). Treatment contains the treatment layout (1,2,3,4,5) and will be the fixed effect of the model, whilst block describes the stratification of this experimental layout according to the random effect "Room".
print(Metadata[1:5,])
genomica(Data = Data,Metadata = Metadata,
Predictors = c('Treatment'),P1_Levels=c('1','2','3','4','5'),
R_Effects=c('Block'),R1_Levels=c('1','2','3','5','6','7'),
Already_Log10_transformed = c('NO'),Folder_Name=c('Test'))
}
