% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomica_one_predictor.R
\name{genomica_one_predictor}
\alias{genomica_one_predictor}
\title{Genomica - analysis with only one predictor}
\usage{
genomica_one_predictor(
  Data,
  Metadata,
  Predictor,
  P_Levels = c(0),
  R_Effects,
  R1_Levels,
  Already_Log10_transformed = c("NO"),
  Folder_Name
)
}
\arguments{
\item{Data}{A data frame with samples as columns and features as rows.}

\item{Metadata}{A data frame with samples as rows.}

\item{Predictor}{example: c('Treatment') - This is the main predictor (independent variable) according to your study design.}

\item{R_Effects}{example: c('Room') - This vector defines the random effects according to your study design.}

\item{R1_Levels}{example: c('Room1,'Room2,Room3,Room4,'Room5,Room6,Room7) - This vector will specify the levels for the random effect, according to your study design.}

\item{Already_Log10_transformed}{Default = c('NO'). If c('YES'), the log10 (n+1) transformation of the features in the object Data will not be performed.}

\item{Folder_Name}{Example: c('Test') - This string will be attached to the folder name in which the results will be generated (e.g.,"Genomica_Output_Test/")}

\item{P_levels}{example: c('Control','Treatment1','Treatment2) - If your main predictor is not a numerical variable, this will set the levels for your main predictor (the model will test the comparisons towards the first term in this vector, such as Control Vs Treatment1 and Control Vs Treatment2 in the example).}
}
\description{
This function will perform a linear mixed model on all the features in your data set, whilst also taking into account of your experimental design (i.e., fixed and random effects). Please see the vignette for more information on how to prepare your data frames for the analysis.
}
\examples{
#Load Data data frame (Data_Demo)
Data<-Data_Demo
#In this example Data is a data frame of 3 taxonomical features (rows) across 42 samples (columns)
print(Data[,1:5])
Metadata<-Metadata_Demo
#in this example Metadata is a data frame with 42 rows (sample IDs) and 5 columns (Sample ID, Treatment, Sex, Room and Block). Treatment contains the treatment layout (Control, T1 or T2) and will be the fixed effect of the model, whilst room describes the stratification of this experimental layout according to the random effect "Room".
print(Metadata[1:5,])
genomica_one_predictor(Data = Data,Metadata = Metadata,
Predictor = c('Treatment'),P_Levels=c('Control','T1','T2'),
R_Effects=c('Room'),R1_Levels=c('Room1','Room2','Room3','Room4','Room5','Room6','Room7'),
Already_Log10_transformed = c('NO'),Folder_Name=c('Test'))
}
