% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomica_two_predictors.R
\name{genomica_two_predictors}
\alias{genomica_two_predictors}
\title{Genomica - analysis with two predictors}
\usage{
genomica_two_predictors(
  Data,
  Metadata,
  Predictor1,
  P1_Levels = c(0),
  Predictor2,
  P2_Levels = c(0),
  R_Effects,
  R1_Levels,
  Already_Log10_transformed = c("NO"),
  Folder_Name
)
}
\arguments{
\item{Data}{A data frame with samples as columns and features as rows.}

\item{Metadata}{A data frame with samples as rows.}

\item{Predictor1}{example: c('Treatment') - This is the first predictor (independent variable) according to your study design.}

\item{Predictor2}{example: c('Sex') - This is the second predictor (independent variable) according to your study design.}

\item{R_Effects}{example: c('Room') - This vector defines the random effects according to your study design.}

\item{R1_Levels}{example: c('Room1,'Room2,Room3,Room4,'Room5,Room6,Room7) - This vector will specify the levels for the random effect, according to your study design.}

\item{Already_Log10_transformed}{Default = c('NO'). If c('YES'), the log10 (n+1) transformation of the features in the object Data will not be performed.}

\item{Folder_Name}{Example: c('Test') - This string will be attached to the folder name in which the results will be generated (e.g.,"Genomica_Output_Two_Predictors_Test/")}

\item{P1_levels}{example: c('Control','Treatment1','Treatment2) - If your main predictor is not a numerical variable, this will set the levels for your main predictor (the model will test the comparisons towards the first term in this vector, such as Control Vs Treatment1 and Control Vs Treatment2 in the example).}

\item{P2_levels}{example: c('Females','Males') - If your second predictor is not a numerical variable, this argument will set the levels for the comparison (the model will test the comparisons towards the first term in this vector, such as Females Vs Males, moreover, the baseline level for the interaction will be in this case Control-Females).}
}
\description{
This function will perform a linear mixed model with fixed effects set as y ~ predictor1 * predictor2. The function will therefore perform multiple hypothesis testing on all the features in your data set, whilst also taking into account of your experimental design (i.e., fixed and random effects). Please see the vignette for more information on how to prepare your data frames for the analysis.
}
\examples{
#Load Data data frame (Data_Demo)
Data<-Data_Demo
#In this example Data is a data frame of 3 taxonomical features (rows) across 42 samples (columns)
print(Data[,1:5])
Metadata<-Metadata_Demo
#in this example Metadata is a data frame with 42 rows (sample IDs) and 3 columns (Sample ID, Treatment and Room). Treatment and sex, with levels c('Control', 'T1','T2') and c('Females','Males) will be the fixed effects for the model, whilst room (describes the stratification of this experimental layout according to the random effect "Room".
print(Metadata[1:5,])
genomica_two_predictors(Data = Data,Metadata = Metadata,
Predictor1= c('Treatment'),
Predictor2=c('Sex'),
P1_Levels=c('Control','T1','T2'),
P2_Levels=c('Female','Male'),
R_Effects=c('Room'),
R1_Levels=c('Room1','Room2','Room3','Room4','Room5','Room6','Room7'),
Already_Log10_transformed = c('No'),Folder_Name=c('Test'))
}
